requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
Flask==2.2.1
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pyOpenSSL==22.0.0
Werkzeug==2.2.1
zipp==3.8.1
